Class and view todo is separated
