+++ src/index.js
@@ -3,4 +3,4 @@ import TodoList from './display_todo.js';
 import { todos } from './array_todo.js';
 
 const todolist = new TodoList(todos);
-return todolist;
+todolist.displayTodos(todos);
